usb: ci13xxx_udc: Add print when Port Reset bit is not set
When the USB device controller detects a USB Reset, the PR bit(PORT RESET)
of PORTSC register is set which is later clear once USB Reset is done.
It is possible that due to interrupt latency the USB software receives
USB BUS Reset Interrupt after the device controller clears the PR bit.
This change adds check and prints message to detect this scenario.

Signed-off-by: Saket Saurabh <ssaurabh@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>